refactor: 토론 유형 토글 컴포넌트로 분리
