+++ src/page/TableListPage/components/Modal/ModalToMakeTable.tsx
@@ -1,4 +1,5 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
+import ToggleForDebateType from './ToggleForDebateType';
 
 interface ModalProps {
   isOpen: boolean;
@@ -6,9 +7,6 @@ interface ModalProps {
 }
 
 export default function ModalToMakeTable({ isOpen, onClose }: ModalProps) {
-  const [isToggleOpen, setIsToggleOpen] = useState(false);
-  const [selectedType, setSelectedType] = useState('의회식 토론 (디폴트)');
-
   useEffect(() => {
     const handleEsc = (e: KeyboardEvent) => {
       if (e.key === 'Escape') {
@@ -21,18 +19,6 @@ export default function ModalToMakeTable({ isOpen, onClose }: ModalProps) {
   }, [onClose]);
 
   if (!isOpen) return null;
-
-  const handleTypeSelect = (type: string) => {
-    setSelectedType(type);
-    setIsToggleOpen(false);
-  };
-
-  // 현재 선택된 값의 반대 옵션을 반환하는 함수
-  const getAlternativeOption = () => {
-    return selectedType === '의회식 토론 (디폴트)'
-      ? '시간 총량제 토론'
-      : '의회식 토론 (디폴트)';
-  };
 
   return (
     <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/50">
@@ -56,22 +42,7 @@ export default function ModalToMakeTable({ isOpen, onClose }: ModalProps) {
           </div>
           <div className="flex items-center justify-center gap-4">
             <h1 className="w-[400px] text-5xl font-bold">토론 유형</h1>
-            <div className="relative">
-              <button
-                onClick={() => setIsToggleOpen(!isToggleOpen)}
-                className="w-[600px] rounded-md bg-neutral-300 p-6 text-center text-3xl font-semibold text-white"
-              >
-                ▼ &nbsp; &nbsp; {selectedType}
-              </button>
-              {isToggleOpen && (
-                <div
-                  onClick={() => handleTypeSelect(getAlternativeOption())}
-                  className="absolute left-0 right-0 top-full mt-2 cursor-pointer rounded-md bg-neutral-300 p-4 text-center text-3xl font-semibold text-white hover:bg-neutral-400"
-                >
-                  {getAlternativeOption()}
-                </div>
-              )}
-            </div>
+            <ToggleForDebateType />
           </div>
         </section>
         <button className="h-[80px] w-full bg-red-200 text-4xl font-semibold transition duration-300 hover:bg-red-300">
